% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAndPlate.R
\name{.collectMeasurements}
\alias{.collectMeasurements}
\title{Collect measurements from a list of plates}
\usage{
.collectMeasurements(plates, order.by = c("columns", "rows"))
}
\arguments{
\item{plates}{A list of matrixes of type character, as generated
by '.collectPlates'.}

\item{order.by}{A character vector, either "columns" or "rows" (or
an abbreviation of those). Defaults to "columns". If
"columns", the column order of the output data.frame is A1,
A2, A3, .... If "rows", the column order of the output
data.frame is A1, B1, C1, ....}
}
\value{
A data.frame with the plate measurements. Each well with
    at least one measurement is represented by a column, each
    plate is represented by a row. Column names are the
    identifiers for the wells.
}
\description{
This function extracts the measurements from plates stored in a
list of matrices of type character, as generated by
'.collectPlates'. Each matrix is assumed to be a plate, with row-
and column-names present. Row-names are assumed to be A, B, ...,
and column-names 1, 2, .... The measurements from each plate are
extracted, and combined in a single data.frame, with one column
per well. Only wells which have at least one data point are
included.
}
\details{
The columns of the data.frame can be ordered by the row or column
of the plate. This can be convenient when replicates are placed in
columns or in rows.
}
