% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAndPlate.R
\name{.collectPlates}
\alias{.collectPlates}
\title{Extract individual plates from an excel file}
\usage{
.collectPlates(filename)
}
\arguments{
\item{filename}{A character variable with the file name of the
excel file.}
}
\value{
A list of matrices of type character. The names of the
    list are the identifiers described above.
}
\description{
This function extracts individual plate data from an excel file
containing readouts of plates. The excel file can contain multiple
sheets and/or multiple plates per sheet. Each plate should have
it's own row- and column-names, with row-names A, B, ..., and
column-names 1, 2, .... Each readout is assumed to be part of a
series, with a numeric identifier in the top-left corner of the
plate (same row as colunm-names and same column as row-names).
}
