% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAndPlate.R
\name{.writeMeasurements}
\alias{.writeMeasurements}
\title{Write collected measurements to a new file}
\usage{
.writeMeasurements(data, filename, format = c("excel", "xlsx", "csv"))
}
\arguments{
\item{data}{A data.frame with the collected measurements, as
generated by '.collectMeasurements'.}

\item{filename}{The name of the output file.}

\item{format}{The character variable specifying the output file
format. Can be "excel" or "xlsx" to output an excel file; or
"csv" to output a csv file.}
}
\value{
Nothing is returned but as a side effect the file
    specified in 'filename' is (over)written.
}
\description{
This function writes the collected measurements, as generated by
'.collectMeasurements' to a new file. This can either be an excel
file or a csv file.
}
