% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAndPlate.R
\name{reformat.plates}
\alias{reformat.plates}
\title{Extract repeated plate measurements from an excel file and order measurements by well}
\usage{
reformat.plates(infile, outfile = NULL, order.by = c("columns", "rows"))
}
\arguments{
\item{infile}{A character variable with the file name of the excel
file containing the plate readouts.}

\item{outfile}{An optional character variable with the file name
of the output file. If present has to end with ".xlsx" (to
output an excel file) or ".csv" (to output a csv file). If no
output file name is given, an excel file is generated using
the same name as the input file, but with '.reformat'
appended.}

\item{order.by}{An optional character vector, either "columns" or
"rows" (or an abbreviation of those). Defaults to
"columns". If "columns", the column order of the output
data.frame is A1, A2, A3, .... If "rows", the column order of
the output data.frame is A1, B1, C1, ....}
}
\value{
Silently returns a data.frame with the plate
    measurements. Each well with at least one measurement is
    represented by a column, each plate is represented by a
    row. Column names are the identifiers for the wells. The rows
    of the data.frame are ordered by the numeric identifier of the
    plates, and the columns (depending on the value of 'order.by')
    either rows or columns of the plate. As a side effect the file
    specified in 'outfile' is (over)written.
}
\description{
This function extracts individual plate data from an excel file
containing readouts of plates. The excel file can contain multiple
sheets and/or multiple plates per sheet. Each plate should have
it's own row- and column-names, with row-names A, B, ..., and
column-names 1, 2, .... Each readout is assumed to be part of a
series, with a numeric identifier in the top-left corner of the
plate (same row as colunm-names and same column as row-names).
}
\details{
The measurements from each plate are extracted, and combined in a
single data.frame, with one column per well. Only wells which have
at least one data point are included.

The columns of the data.frame can be ordered by the row or column
of the plate. This can be convenient when replicates are placed in
columns or in rows.

The collected measurements are returned as a data.frame and
written to a new file (either excel or csv).
}
