\name{timeAndPlate-package}
\alias{timeAndPlate-package}
\alias{timeAndPlate}
\docType{package}
\title{\packageTitle{timeAndPlate}}
\description{\packageDescription{timeAndPlate}}
\section{Package Content}{\packageIndices{timeAndPlate}}
\author{\packageAuthor{timeAndPlate}}
\section{Maintainer}{\packageMaintainer{timeAndPlate}}
\keyword{package}
